refactor(core): extract the streaming model fetch
Move SHA-256 staging, the size cap, and atomic rename into
download/fetch.rs. Presence checks and ensure_* stay in download.rs.
