+++ CHANGELOG.md
@@ -27,7 +27,8 @@ were released without a git tag, so their headings carry no compare link.
 - Nightly AddressSanitizer job over the session pool (alongside Miri + TSAN).
 - `Engine` impl split into `engine/{config,load,stream,transcribe,infer}.rs`.
   Public API is unchanged.
-- `model` download/identity split into `model/{progress,variant,download}.rs`.
+- `model` download/identity split into `model/{progress,variant,download}.rs`;
+  the streaming SHA-256 fetch lives in `download/fetch.rs`.
   Public `gigastt_core::model::*` paths are unchanged. CLI parse tests live
   next to `main.rs` split by command (`cli_tests/{serve,download,transcribe,
   batch,parse}.rs`); bind-gate tests live next to `serve.rs`.
